requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
filelock==3.28.0
fsspec[http]==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.10.2
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.1
prompt-toolkit==3.0.52
pygments==2.20.0
pyyaml==6.0.3
rich==15.0.0
shellingham==1.5.4
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
wcwidth==0.6.0
# Below are dependencies of lm-eval
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
chardet==5.2.0
charset-normalizer==3.4.7
colorama==0.4.6
dataproperty==1.1.0
datasets==4.8.4
dill==0.4.1
evaluate==0.4.6
frozenlist==1.8.0
jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
lm-eval==0.4.11
lxml==6.0.4
markupsafe==3.0.3
mbstrdecoder==1.1.4
more-itertools==11.0.2
multidict==6.7.1
multiprocess==0.70.19
nltk==3.9.4
numpy==2.4.4
pandas==3.0.2
pathvalidate==3.3.1
portalocker==3.2.0
propcache==0.4.1
pyarrow==23.0.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
regex==2026.4.4
requests==2.33.1
rouge-score==0.1.2
sacrebleu==2.6.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sqlitedict==2.1.0
tabledata==1.3.4
tabulate==0.10.0
tcolorpy==0.1.7
threadpoolctl==3.6.0
typepy[datetime]==1.3.4
urllib3==2.6.3
word2number==1.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
